clean-up file structure
